Clean up Kubelet Client/Config Logic
This brings KubeletClientConfig and KubeletClient more in line with
rest.Config and RESTClient -- you call `KubeletClientFor(config)`, and
it produces a new `KubeletInterface` that just contains the needed bits
of the config in its own fields.

It also simplifies the transport construction code to use standard
methods, and removes unused fields.

The summary provider now also accepts a kubelet client directly,
instead of trying to construct it itself.  This centralizes client
construction.