Rework internal metrics pipeline
This *greatly* simplifies the internal metrics pipeline.

The last of the processors, the rate calculator, is removed.
We can get CPU usage rate directly from the summary API.

The metric API are now served from a pair of "provider" interfaces,
which abstract over the process of fetching metrics from the actual
underlying store.

These providers are backed by a very simple metrics sink, which stores
only the most recent batch of data, and is structed by nodes and pods,
for easy lookup.

Internal data is now structure in a way that is conducive to serving the
resource metrics API -- there are node metric points, and there are pod
metric points, with each pod containing the metric points for its
containers.

Since this is how the summary API sends data, the summary API code is
also simplified greatly.  We now also only save the data we need, namely
CPU and memory usage.

This also strips away a lot of unneeded complexity from the manager,
turning it into a straightfoward ticker-based loop.