Refactor Makefile and Update Dockerfile
This refactors the makefile, cleaning it up a bit, removing some
unused logic, and moving build information gathering into a sub-makefile.

This also updates the Dockerfile, to no longer require CA certs, and
updates the deployment to allow images from scratch or read-only
containers by mounting a volume for `/tmp`.