Switch node address resolution to match kube
This switches over the node address resolution to match the default in
the Kubernetes API server used for the nodes/proxy subresource.

Namely, it prefers explicitly set hostname first, then internal
addresses, then external, preferring DNS names to IP addresses.

This means that if resolution doesn't do what you want, at least it
breaks the same was as the main Kubernetes API server.

We also expose a similar flag at the command line to Kubernetes for
configuring this ordering.