Serve OpenAPI scheme
This updates our server code to actually serve an OpenAPI schema, so
that the aggregator will stop logspamming (and people can discover
useful things about the objects we return).

The generated openapi files are not checked in.  Instead, we
force-include openapi-gen as a dep, and generate openapi definitions at
build time.