Allow plumbing windows through from provider
This allows plumbing the window through from the provider.  In the case
of the sink provider, it could allow us to provide more accurate windows
(either the scrape period or the actual window, if available), and in
external providers it should allow them to provide acurate windows
as well from their calculations.

It also ensures that the *earliest* timestamp is always used when
aggregating timestamps together, so that we can know if a given
timestamp could be tainted by things like pod initialization.