Hardcode sink provider's window to better guess
While the summary API doesn't actually expose the window used to
calculate instantaneous CPU usage (it *could*, but it doesn't), we can
guess at the max by taking the normal max housekeeping interval (15s),
and multiplying it by two to account for jitter, which can up-to-double
the actual housekeeping interval.  In normal operation, it's probably
anywhere between 10-20s.