Don't discard node results on error
It's possible for the summary provider to return errors and partial
results.  This is explicitly listed in the documentation for the Source
interface.  We shouldn't discard partial results just because they're
partial, otherwise we'll end up discarding a node's worth of results
because of an error collecting for one pod.