Maintain the same ordering invariant that the Kube API server does
For consistency with the rest of the API ecosystem, we should have
a consistent default order with the upstream APIs for nodes and
pods. The cost is relatively small for nodes and for viewing all
pods should be less expensive than populating the pod list with
metrics for all pods.