Set pod priority to system-cluster-critical
As part of autoscaling pipeline Metrics Server is expected to be system
critical in clusters that rely on autoscaling. system-cluster-critical
is one of the higher priorities built in kubernetes. Running in higher
priority means that metrics-server has priority when scheduling and lack
of capacity in cluster could cause pods with lower priority to be
preempted.