test: force metrics initialization
Metrics created with the Kubernetes library are lazily instantiated.
Their initialization is delayed until registration. Till then, metrics
are neither describable nor collectible.
In the tests we are using the testutil CollectAndCompare function to
gather the metrics and verify them, and it requires a valid Collector.
To force its initialization we need to call the Create method
beforehand.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>