Watch only metadata for pods
Metrics Server requires pod metadata to filter pods by label
selectors. Pod Spec and Status are not used, thus we can replace normal
informer with metadata informer which avoids parsing and storing them in
memory. Only field that from status that is used if to filter if pod is
running, but this can be done on as field selector.

This reduces memory usage by over 40% and removes memory overhead of
watching pods in Terminated state.