Create TOC for KNOWN_ISSUES.md and FAQ.md
This commit adds a table of contents to the KNOW_ISSUES and the FAQ.
In addition to that, it automates the verification and update of these
TOCs via the verify-toc and update-toc Makefile rules. These new rules
have been added to the verify and update rules, meaning that verify-toc
is now part of the presubmit tests.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>