pkg/storage: add unit tests for cumulative metrics
With the storage rewrite to use cumulative metrics, a couple of new test
cases needs to be added since metrics-server now requires two scrapes to
compute the CPU usage of Pods and Nodes. The new tests cover adding
pods/nodes that were not in the previous batch of metrics and removing
pods/nodes that were in the previous batch. In both cases, we want
metrics-server to return nil metrics for the requested ressource since
it doesn't have enough data to compute the CPU usage. Another test was
added to verify that if two scrapes yields the same metrics points for a
pod/node, the store isn't update with the new data. This makes sure that
we do not compute the CPU usage based on the same metrics points.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>