pkg/storage: store previous points only if present
Previously, we used to create entries in the previous metrics points map
even when there was none. The new entry was set to the default value of
a metric point and we had to compare its timestamp to zero to know if it
existed or not. This made checking for presence of previous metrics
points unatural, so the storing was improve to only store known metrics
points.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>