*: add HA manifests for different compatibility
Going forward, we will use the high-availability component as the base
to build compatible configuration on top. Then each changes impacting
compatibility will have to introduce a new component that will patch the
main one.

This is done to make sure that we keep the same assets semantic across
versions and that we do not make incompatible changes that would break
compatibility with older version of Kubernetes.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>