.github: do not publish assets with helm chart
We used to publish metrics-server's assets alongside the helm chart
although they are not useful to it because the helm chart release was
tagged as "latest" and so if we wanted users to get the latest assets we
needed to copy them to the helm release. Since helm releases are not
tagged as latest anymore, it is not necessary do continue doing that.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>